feat: Adding Book and BookShelf components
Adding the Book and BookShelf components and refactoring the App
component to use BookShelf, and BookShelf component to use Book component.
Book component is responsible to draw one book with the state representing
the current shelf the book on, the component recieve book
and change listener as props, and propagate the changing shelf event
to bookshelf. BookShelf represents a list of books on a certain shelf.